requirements.txt
Pillow
numpy
keras
flyai
